feat: move vroom functions to own package
