+++ internal/storage/routing/valhalla/repo.go
@@ -10,17 +10,13 @@ import (
 	"github.com/green-ecolution/green-ecolution-backend/internal/config"
 	"github.com/green-ecolution/green-ecolution-backend/internal/entities"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage"
-	"github.com/green-ecolution/green-ecolution-backend/internal/storage/routing/openrouteservice/vroom"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/routing/valhalla/valhalla"
+	"github.com/green-ecolution/green-ecolution-backend/internal/storage/routing/vroom"
 	"github.com/green-ecolution/green-ecolution-backend/internal/utils"
 )
 
 // validate is RouteRepo implements storage.RoutingRepository
 var _ storage.RoutingRepository = (*RouteRepo)(nil)
-
-const (
-	treeAmount int32 = 40 // how much water does a cluster need
-)
 
 type RouteRepoConfig struct {
 	routing config.RoutingConfig
@@ -44,6 +40,9 @@ func NewRouteRepo(cfg *RouteRepoConfig) (*RouteRepo, error) {
 
 	vroomClient := vroom.NewVroomClient(
 		vroom.WithHostURL(vroomURL),
+		vroom.WithStartPoint(cfg.routing.StartPoint),
+		vroom.WithEndPoint(cfg.routing.EndPoint),
+		vroom.WithWateringPoint(cfg.routing.WateringPoint),
 	)
 	valhalllaClient := valhalla.NewValhallaClient(
 		valhalla.WithHostURL(valhallaURL),
@@ -75,7 +74,7 @@ func (r *RouteRepo) GenerateRawGpxRoute(ctx context.Context, vehicle *entities.V
 }
 
 func (r *RouteRepo) prepareRoute(ctx context.Context, vehicle *entities.Vehicle, clusters []*entities.TreeCluster) (*valhalla.DirectionRequest, error) {
-	optimizedRoutes, err := r.optimizeRoute(ctx, vehicle, clusters)
+	optimizedRoutes, err := r.vroom.OptimizeRoute(ctx, vehicle, clusters)
 	if err != nil {
 		slog.Error("failed to optimize route", "error", err)
 		return nil, err
@@ -131,84 +130,4 @@ func (r *RouteRepo) prepareRoute(ctx context.Context, vehicle *entities.Vehicle,
 		Costing:        "truck",
 		CostingOptions: costingOpts,
 	}, nil
-}
-
-func (r *RouteRepo) optimizeRoute(ctx context.Context, vehicle *entities.Vehicle, cluster []*entities.TreeCluster) (*vroom.VroomResponse, error) {
-	vroomVehicle, err := r.toVroomVehicle(vehicle)
-	if err != nil {
-		if errors.Is(err, storage.ErrUnknownVehicleType) {
-			slog.Error("unknown vehicle type. please specify vehicle type", "error", err, "vehicle_type", vehicle.Type)
-		}
-
-		return nil, err
-	}
-
-	shipments := r.toVroomShipments(cluster)
-	req := &vroom.VroomReq{
-		Vehicles:  []vroom.VroomVehicle{*vroomVehicle},
-		Shipments: shipments,
-	}
-
-	resp, err := r.vroom.Send(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	return resp, nil
-}
-
-func (r *RouteRepo) toVroomShipments(cluster []*entities.TreeCluster) []vroom.VroomShipments {
-	// ignore tree cluster with empty coordinates
-	filteredClusters := utils.Filter(cluster, func(c *entities.TreeCluster) bool {
-		return c.Longitude != nil && c.Latitude != nil
-	})
-
-	nextID := int32(0)
-	return utils.Map(filteredClusters, func(c *entities.TreeCluster) vroom.VroomShipments {
-		shipment := vroom.VroomShipments{
-			Amount: []int32{treeAmount},
-			Pickup: vroom.VroomShipmentStep{
-				ID:       nextID,
-				Location: r.cfg.routing.WateringPoint,
-			},
-			Delivery: vroom.VroomShipmentStep{
-				Description: c.Name,
-				ID:          nextID + 1,
-				Location:    []float64{*c.Longitude, *c.Latitude},
-			},
-		}
-
-		nextID += 2
-		return shipment
-	})
-}
-
-func (r *RouteRepo) toVroomVehicle(vehicle *entities.Vehicle) (*vroom.VroomVehicle, error) {
-	vehicleType, err := r.toOrsVehicleType(vehicle.Type)
-	if err != nil {
-		return nil, err
-	}
-
-	return &vroom.VroomVehicle{
-		ID:          vehicle.ID,
-		Description: vehicle.Description,
-		Profile:     vehicleType,
-		Start:       r.cfg.routing.StartPoint,
-		End:         r.cfg.routing.EndPoint,
-		Capacity:    []int32{int32(vehicle.WaterCapacity)}, // vroom don't accept floats
-	}, nil
-}
-
-func (r *RouteRepo) toOrsVehicleType(_ entities.VehicleType) (string, error) {
-	return "driving-car", nil
-	// switch vecType {
-	// case entities.VehicleTypeTrailer:
-	// 	return "driving-car", nil
-
-	// case entities.VehicleTypeTransporter:
-	// 	return "driving-hgv", nil
-
-	// default:
-	// 	return "", storage.ErrUnknownVehicleType
-	// }
 }
