+++ internal/storage/routing/vroom/entities.go
