+++ main.go
@@ -30,6 +30,7 @@ import (
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/auth"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/local"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/postgres"
+	_ "github.com/green-ecolution/green-ecolution-backend/internal/storage/routing/openrouteservice"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/routing/valhalla"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/s3"
 	"github.com/green-ecolution/green-ecolution-backend/internal/worker"
@@ -132,11 +133,14 @@ func initializeRepositories(ctx context.Context, cfg *config.Config) (*storage.R
 		return nil, nil, fmt.Errorf("error creating local repository: %w", err)
 	}
 
-	keycloakRepo := auth.NewRepository(&cfg.IdentityAuth)
+	// can be switched between ors and valhalla
+	// routingRepo, err := openrouteservice.NewRepository(cfg)
 	routingRepo, err := valhalla.NewRepository(cfg)
 	if err != nil {
 		panic(err)
 	}
+
+	keycloakRepo := auth.NewRepository(&cfg.IdentityAuth)
 
 	s3Repos, err := s3.NewRepository(cfg)
 	if err != nil {
